# ⚡️ Avoid creating temporary expression in matrix multiplications