--- a/src/quad_expr.jl
+++ b/src/quad_expr.jl
@@ -126,55 +126,64 @@ function Base.eltype(qti::QuadTermIterator{GenericQuadExpr{C, V}}
     return Tuple{C, V, V}
 end
 
-# With one factor
+# With one factor.
 
 function add_to_expression!(quad::GenericQuadExpr{C}, other::C) where C
     return add_to_expression!(quad.aff, other)
 end
 
-function add_to_expression!(q::GenericQuadExpr{T,S}, other::GenericAffExpr{T,S}) where {T,S}
+function add_to_expression!(q::GenericQuadExpr{T,S},
+                            other::GenericAffExpr{T,S}) where {T,S}
     add_to_expression!(q.aff, other)
     return q
 end
 
-function add_to_expression!(q::GenericQuadExpr{T,S}, other::GenericQuadExpr{T,S}) where {T,S}
+function add_to_expression!(q::GenericQuadExpr{T,S},
+                            other::GenericQuadExpr{T,S}) where {T,S}
     merge!(+, q.terms, other.terms)
     add_to_expression!(q.aff, other.aff)
     return q
 end
 
-# With two factors
+# With two factors.
 
 function add_to_expression!(quad::GenericQuadExpr{C, V},
-                            new_coef::Union{C, Real},
-                            new_aff::Union{V, GenericAffExpr{C,V}}) where {C,V}
+                            new_coef::Real,
+                            new_var::V) where {C,V}
+    add_to_expression!(quad.aff, new_coef, new_var)
+    return quad
+end
+
+function add_to_expression!(quad::GenericQuadExpr{C, V},
+                            new_var::Union{V, GenericAffExpr{C, V}},
+                            new_coef::Real) where {C,V}
+    return add_to_expression!(quad, new_coef, new_var)
+end
+
+function add_to_expression!(quad::GenericQuadExpr{C},
+                            new_coef::Real,
+                            new_aff::GenericAffExpr{C}) where {C}
     add_to_expression!(quad.aff, new_coef, new_aff)
     return quad
 end
 
-function add_to_expression!(quad::GenericQuadExpr{C, V},
-                            new_aff::Union{V, GenericAffExpr{C,V}},
-                            new_coef::Union{C, Real}) where {C,V}
-    add_to_expression!(quad.aff, new_aff, new_coef)
-    return quad
-end
-
-function add_to_expression!(quad::GenericQuadExpr{C,V}, coef::Union{C, Real},
-                            other::GenericQuadExpr{C,V}) where {C,V}
+function add_to_expression!(quad::GenericQuadExpr{C, V}, coef::Real,
+                            other::GenericQuadExpr{C, V}) where {C, V}
     for (key, term_coef) in other.terms
         _add_or_set!(quad.terms, key, coef * term_coef)
     end
     return add_to_expression!(quad, coef, other.aff)
 end
 
-function add_to_expression!(quad::GenericQuadExpr{C,V},
-                            other::GenericQuadExpr{C,V},
-                            coef::Union{C, Real}) where {C,V}
+function add_to_expression!(quad::GenericQuadExpr{C, V},
+                            other::GenericQuadExpr{C, V},
+                            coef::Real) where {C, V}
     return add_to_expression!(quad, coef, other)
 end
 
-function add_to_expression!(quad::GenericQuadExpr{C,V},
-                            var_1::V, var_2::V) where {C,V}
+function add_to_expression!(quad::GenericQuadExpr{C},
+                            var_1::AbstractVariableRef,
+                            var_2::AbstractVariableRef) where {C}
     return add_to_expression!(quad, one(C), var_1, var_2)
 end
 
@@ -238,9 +247,10 @@ function add_to_expression!(quad::GenericQuadExpr{C,V},
     return quad
 end
 
-# With three factor
+# With three factors.
 
-function add_to_expression!(quad::GenericQuadExpr{C,V}, new_coef::C, new_var1::V, new_var2::V) where {C,V}
+function add_to_expression!(quad::GenericQuadExpr{C,V}, new_coef::C,
+                            new_var1::V, new_var2::V) where {C,V}
     # Node: OrderedDict updates the *key* as well. That is, if there was a
     # previous value for UnorderedPair(new_var2, new_var1), it's key will now be
     # UnorderedPair(new_var1, new_var2) (because these are defined as equal).