With the benchmark in https://github.com/JuliaOpt/JuMP.jl/issues/1403#issuecomment-463857803:
Before:
```
julia> @time x*A;
0.559247 seconds (8.21 M allocations: 668.261 MiB, 40.98% gc time)

julia> @time A*x*A;
33.784054 seconds (14.60 M allocations: 9.227 GiB, 51.83% gc time)
```
After:
```
julia> @time x*A;
0.153061 seconds (210.03 k allocations: 73.168 MiB)

julia> @time A*x*A;
16.037227 seconds (630.06 k allocations: 6.126 GiB, 14.55% gc time)
```