@mlubin 2019-02-15
Could we add this case to `test/perf`?